+++ CHANGELOG.md
@@ -2491,3 +2491,4 @@
 - [2026-07-02 18:52] chore(cache): update dependencies and cleanup unused imports in cache
 - [2026-07-02 19:47] perf(cache): apply RepaintBoundary on animated cache widgets
 - [2026-07-02 19:50] test(perf): add unit tests for perf business logic
+- [2026-07-02 20:14] fix(cache): prevent double tap duplicate action in cache